requirements.txt
numpy
sklearn
matplotlib
seaborn
pandas
umap-learn
tqdm
plotly
streamlit